% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{enps_to_entrez}
\alias{enps_to_entrez}
\title{Mapping from Ensembl to Entrez}
\usage{
enps_to_entrez(ens_proteins, mart)
}
\arguments{
\item{ens_proteins}{list of Ensembl IDs}

\item{mart}{a biomaRt object for the given species}
}
\value{
a data.frame representing the mapping
}
\description{
Maps a list of protein IDs in the Ensmbl format to the Entrez naming system
}
