% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{expand_seeds}
\alias{expand_seeds}
\title{Expansion of the seed network (fromt he genomic context)}
\usage{
expand_seeds(genomic_context, ppi_network, ensp_to_ensg, strict_proteins,
  strict_connectors = TRUE)
}
\arguments{
\item{genomic_context}{a two column data.fram produced by \code{\link[LErNet]{LErNet.get_genomic_context}}}

\item{ppi_network}{a two column data.frame representing PPI network edges (see also \code{\link[LErNet]{LErNet.get_stringdb}} )}

\item{ensp_to_ensg}{a two column data.frame for mapping proteins to their producer genes (see also \code{\link[LErNet]{LErNet.get_stringdb}} )}

\item{strict_proteins}{a list of proteins}

\item{strict_connectors}{if \code{TRUE} connectors can onyl be choosen from the \code{strict_proteins} list}
}
\value{
a list
\describe{
  \item{network_components}{a list of connected components of the resultant expanded network. Each compoent is a list of proteins.}
  \item{network_seeds}{list of seed proteins that have succefully been mapped to the PPI network.}
}
}
\description{
Expands with connectors the network formed by seed proteins,
that are the producs fo the genes int he genomic context,
by the expasion algorithm.
Connectors are neighbors of selected proteins in the input PPI network.
}
\examples{

}
