% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arenaidb.R
\name{get_arenaidb_diseases}
\alias{get_arenaidb_diseases}
\title{Retriving the disease's section of Arena-idb}
\usage{
get_arenaidb_diseases()
}
\value{
arenaidb a data.frame where for each lncrna is specified the diseases in which it can be involved
}
\description{
Retriving the disease's section of Arena-idb
}
\examples{
diseases <- get_arenaidb_diseases()

}
