% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arenaidb.R
\name{get_arenaidb_gene_aliases}
\alias{get_arenaidb_gene_aliases}
\title{Retriving the alias of genes from Arena-idb}
\usage{
get_arenaidb_gene_aliases()
}
\description{
Retriving the alias of genes from Arena-idb
}
\examples{
gene_aliases <- get_arenaidb_gene_aliases()

}
