% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arenaidb.R
\name{get_arenaidb_interactions}
\alias{get_arenaidb_interactions}
\title{Retriving the interactions involving lncrnas}
\usage{
get_arenaidb_interactions()
}
\value{
arenaidb a data.frame containing in the first two columns the informations about the lncrnas and in the following those about the interactions
}
\description{
Retriving the interactions involving lncrnas
}
\examples{
interactions <- get_arenaidb_interactions()

}
