% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arenaidb.R
\name{get_arenaidb_ncrnas}
\alias{get_arenaidb_ncrnas}
\title{Retriving the lncrna's section of Arena-idb}
\usage{
get_arenaidb_ncrnas()
}
\value{
arenaidb a data.frame where for each ncrna are specified all the informations
}
\description{
Retriving the lncrna's section of Arena-idb
}
\examples{
ncrnas <- get_arenaidb_ncrnas()

}
