% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arenaidb.R
\name{get_arenaidb_ncrnas_aliases}
\alias{get_arenaidb_ncrnas_aliases}
\title{Retriving the aliases of lncrnas from Arena-idb}
\usage{
get_arenaidb_ncrnas_aliases()
}
\description{
Retriving the aliases of lncrnas from Arena-idb
}
