% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{get_connected_components}
\alias{get_connected_components}
\title{Extraction of the connected components of the network}
\usage{
get_connected_components(net)
}
\arguments{
\item{net}{a data.frame reporting the edges of the network}
}
\value{
components a list of all connected proteins
}
\description{
Extraction of the connected components of the network
}
\examples{

}
