% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{get_stringdb}
\alias{get_stringdb}
\title{Retrieving of infomraiton fron the STRING database}
\usage{
get_stringdb(stringdb_tax = 9606, stringdb_thr = 900, mart)
}
\arguments{
\item{stringdb_tax}{taxa of the species}

\item{stringdb_thr}{threshold to be applied to the score on the edges of the PPI}

\item{mart}{a biomaRt object for mapping proteins to producer genes (Ensembl IDs).}
}
\value{
a list
\describe{
  \item{\code{ppi_network}}{a two columns data.frame representing the PPI network by listing its edges.}
  \item{\code{ensp_to_ensg}}{a two columns data.frame reporting for each protein the corresponding gene (Ensembl IDs)}
}
}
\description{
Retrieves the PPI network formt he STRING database via the STRINGdb.
STRINGdb often onyl associates a primary product to a gene,
thus other products are not reported.
The function also returns the proteins associated to each gene within the STRING database.
}
\examples{
library(biomaRt)
stringdb_tax = 9606
stringdb_thr = 900
mart = useMart(biomart = "ensembl", dataset = "hsapiens_gene_ensembl")
ret <- LErNet.get_stringdb( stringdb_tax = stringdb_tax, stringdb_thr = stringdb_thr, mart = mart)
ppi_network <- ret[["ppi_network"]]
ensp_to_ensg <- ret[["ensp_to_ensg"]]

}
