% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet_011.R
\name{get_stringdb_bc_011}
\alias{get_stringdb_bc_011}
\title{Retrieving of information from the STRING database}
\usage{
get_stringdb_bc_011(stringdb_tax = 9606, stringdb_thr = 900, mart)
}
\arguments{
\item{stringdb_tax}{taxa of the species}

\item{stringdb_thr}{threshold to be applied to the score on the edges of the PPI}

\item{mart}{a biomaRt object for mapping proteins to producer genes (Ensembl IDs).}
}
\value{
a list
\describe{
  \item{\code{ppi_network}}{a two columns data.frame representing the PPI network by listing its edges.}
  \item{\code{ensp_to_ensg}}{a two columns data.frame reporting for each protein the corresponding gene (Ensembl IDs)}
}
}
\description{
Retrieves the PPI network form the STRING database via the STRINGdb.
STRINGdb often only associates a primary product to a gene,
thus other products are not reported.
The function also returns the proteins associated to each gene within the STRING database.
}
