% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arenaidb.R
\name{search_in_arenaidb}
\alias{search_in_arenaidb}
\title{Retriving the informations about the aliases and the interactions of lncrna filtering by genes and proteins}
\usage{
search_in_arenaidb(
  lncrna_ensembl_ids,
  pcgene_ensembl_ids = NULL,
  protein_ensembl_ids = NULL,
  protein_symbols = NULL
)
}
\arguments{
\item{lncrna_ensembl_ids}{a list of lncrnas in ensembl id format}

\item{pcgene_ensembl_ids}{a list of protein coding genes in ensembl id format}

\item{protein_ensembl_ids}{a list of protein in ensembl id format}

\item{protein_symbols}{a list of protein in symbol id format}
}
\value{
return_value a list of three data.frame: alias_mapping, gene_icontext and protein_icontext
}
\description{
Retriving the informations about the aliases and the interactions of lncrna filtering by genes and proteins
}
\examples{

}
