% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{visualize}
\alias{visualize}
\title{Visualization of the expanded network}
\usage{
visualize(
  lncrna_context,
  de_proteins,
  network_seeds,
  expanded_elements,
  bg_ppi_network,
  labels
)
}
\arguments{
\item{lncrna_context}{the genomic and interaction context of the lncRNAs (see \code{\link[LErNet]{get_genomic_context}}) and \code{\link[LErNet]{search_in_arenaidb}}}

\item{de_proteins}{the complete vector of input proteins,that are of interest for the study}

\item{network_seeds}{the list of seed proteins}

\item{expanded_elements}{the list of proteins (see also \code{\link[LErNet]{expand_seeds}} ) that must be visualized}

\item{bg_ppi_network}{a two column data.frame representing PPI network edges (see also \code{\link[LErNet]{get_stringdb}} )}

\item{labels}{a two column data.frame representing all the labels}
}
\value{
visualizes the network in the Viewer window
}
\description{
Visualizes the expanded network,
composed by seed proteins and connectors.
LncRNA are added together with extra edges in order to report the lncrna context.
LncRNAs are linked to the proteins that are products of their lncrna context.
}
