% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet_011.R
\name{enrich_001}
\alias{enrich_001}
\title{Functional enrichment}
\usage{
enrich_001(ens_proteins, organism, mart)
}
\arguments{
\item{ens_proteins}{list of proteins, in Ensembl format, for which to compute the enrichment}

\item{mart}{a biomaRt object of the given species need for the conversion from  Ensembl to Entrez IDs}

\item{oganism}{oganism name (see \code{\link[ReactomePA]{enrichPathway}})}
}
\value{
a ReactomePA result object.
}
\description{
Computes functional enrichment of a given set of proteins via the ReactomePA package.
}
\examples{

}
