% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{expand_seeds}
\alias{expand_seeds}
\title{Expansion of the seed network}
\usage{
expand_seeds(seeds, ppi_network, de_proteins = NULL)
}
\arguments{
\item{seeds}{a list produced by \code{\link[LErNet]{get_genomic_context}} and \code{\link[LErNet]{search_in_arenaidb}}}

\item{ppi_network}{a two column data.frame representing PPI network edges (see also \code{\link[LErNet]{get_stringdb}} )}

\item{de_proteins}{a vector of stricted proteins. If empty, the complete set of neighbors is extracted otherwise only pcgenes in this set are taken into account.}
}
\value{
a list
\describe{
  \item{out_components}{a list of connected components of the resultant expanded network. Each compoent is a list of proteins.}
}
}
\description{
Expands with connectors the network formed by seed proteins,
that are the produtcs of the genes in the genomic context and in the interaction context,
by the expansion algorithm.
Connectors are neighbors of selected proteins in the input PPI network.
}
\examples{

}
