% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{get_genomic_context}
\alias{get_genomic_context}
\title{Extraction of the genomic context}
\usage{
get_genomic_context(positions, lncgenes, pcgenes = NULL, max_window = 1e+05)
}
\arguments{
\item{positions}{a data.frame reporting genomic positions by \code{\link[LErNet]{load_gtf}}. Columns are \code{id type seqname start end}. It may contain features into listed in \code{lncgenes} and \code{pcgenes}}

\item{lncgenes}{a list of lncRNA genes}

\item{pcgenes}{a list of protein-coding genes that are of interest for the study. If empty, the complete set of neighbors is extracted otherwise only pcgenes in this set are taken into account.}

\item{max_window}{the maximum size of the genomic range}
}
\value{
a two column data.frame reporting neighborhood information. The first column gives lncRNAs and the second column gives their associated neighbors.
}
\description{
Retrieves the genomic context of input lncRNAs.
The genomic context is defined as the set of protin coding genes
that resides within a given range.
}
