% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{get_stringdb}
\alias{get_stringdb}
\title{Retrieving of information from the STRING database}
\usage{
get_stringdb(stringdb_tax = 9606, stringdb_thr = 900)
}
\arguments{
\item{stringdb_tax}{taxa of the species}

\item{stringdb_thr}{threshold to be applied to the score on the edges of the PPI}
}
\value{
a list
\describe{
  \item{\code{ppi}}{a two columns data.frame representing the PPI network by listing its edges.}
}
}
\description{
Retrieves the PPI network form the STRING database via the STRINGdb.
STRINGdb often only associates a primary product to a gene,
thus other products are not reported.
The function also returns the proteins associated to each gene within the STRING database.
}
\examples{

}
