% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LErNet.R
\name{visualize}
\alias{visualize}
\title{Visualization of the expanded network}
\usage{
visualize(lncgenes, genomic_context, ensp_to_ensg, input_proteins,
  network_seeds, ppi_network, expanded_elements, mart,
  mart_symbol_column = NULL)
}
\arguments{
\item{lncgenes}{a list of lncRNA genes}

\item{genomic_context}{the genomc context of the lncRNAs (see \code{\link[LErNet]{get_genomic_context}})}

\item{ensp_to_ensg}{a two column data.frame for mapping proteins to their producer genes (see also \code{\link[LErNet]{get_stringdb}} )}

\item{input_proteins}{the complete list of input proteins,that are of interest for the study}

\item{network_seeds}{the lsit of seed protieins}

\item{ppi_network}{a two column data.frame representing PPI network edges (see also \code{\link[LErNet]{get_stringdb}} )}

\item{expanded_elements}{the list of proteins that must be visualized}

\item{mart}{a biomaRt ojbect used to visualize symbols instead of Ensembl IDs}

\item{mart_symbol_column}{column of the biomaRt ojbect fomr wich symbols are retrieved}
}
\value{
visualizes the network in the Viewer window
}
\description{
Visualizes the expanded network,
composed by seed proteins and connectors.
LncRNA are added together with extra edges in order to report the genomic context.
LncRNAs are linked to the proteins that are products of their genomic context.
}
\examples{

}
